rename: utils 구조 개선
