+++ src/apis/auth/postLogout.ts
@@ -1,7 +1,7 @@
 import { axiosInstance } from '@/apis/axios/axios';
 import type { LogoutResponse } from '@/types/auth/logout';
 import { useMutation } from '@tanstack/react-query';
-import { getRefreshToken } from '@/utils/auth/authStorage';
+import { getRefreshToken } from '@/utils/authStorage';
 
 export const postLogout = async (): Promise<LogoutResponse> => {
   const refreshToken = getRefreshToken();
