+++ src/hooks/useGroupedTags.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useGetTags } from '@/apis/tag/getTags';
-import { splitTags, type TagGroups } from '@/utils/tag/splitTags';
+import { splitTags, type TagGroups } from '@/utils/splitTags';
 
 /**
  * 태그 목록 조회 및 온보딩 그룹별 분류 훅
