+++ src/utils/finalizeLogin.ts
@@ -1,4 +1,4 @@
-import { setAuthTokens } from '@/utils/auth/authStorage';
+import { setAuthTokens } from '@/utils/authStorage';
 import { getUserProfile } from '@/apis/mypage/getUserProfile';
 import { queryKey } from '@/constants/queryKey';
 import type { QueryClient } from '@tanstack/react-query';
